+++ flink-connector/src/test/java/com/datastrato/gravitino/flink/connector/integration/test/hive/FlinkHiveCatalogIT.java
@@ -18,19 +18,11 @@
 package com.datastrato.gravitino.flink.connector.integration.test.hive;
 
 import static com.datastrato.gravitino.catalog.hive.HiveCatalogPropertiesMeta.METASTORE_URIS;
-import static com.datastrato.gravitino.flink.connector.integration.test.utils.TestUtils.assertColumns;
-import static com.datastrato.gravitino.flink.connector.integration.test.utils.TestUtils.toFlinkPhysicalColumn;
-import static com.datastrato.gravitino.rel.expressions.transforms.Transforms.EMPTY_TRANSFORM;
 
-import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.flink.connector.PropertiesConverter;
 import com.datastrato.gravitino.flink.connector.hive.GravitinoHiveCatalog;
 import com.datastrato.gravitino.flink.connector.hive.GravitinoHiveCatalogFactoryOptions;
 import com.datastrato.gravitino.flink.connector.integration.test.FlinkCommonIT;
-import com.datastrato.gravitino.flink.connector.integration.test.utils.TestUtils;
-import com.datastrato.gravitino.rel.Column;
-import com.datastrato.gravitino.rel.Table;
-import com.datastrato.gravitino.rel.types.Types;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import java.util.Arrays;
@@ -39,19 +31,12 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import jline.internal.Preconditions;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.table.api.DataTypes;
-import org.apache.flink.table.api.ResultKind;
 import org.apache.flink.table.api.TableResult;
 import org.apache.flink.table.api.ValidationException;
 import org.apache.flink.table.catalog.Catalog;
-import org.apache.flink.table.catalog.CatalogBaseTable;
 import org.apache.flink.table.catalog.CatalogDescriptor;
-import org.apache.flink.table.catalog.CatalogTable;
 import org.apache.flink.table.catalog.CommonCatalogOptions;
-import org.apache.flink.table.catalog.ObjectPath;
-import org.apache.flink.table.catalog.exceptions.TableNotExistException;
 import org.apache.flink.table.catalog.hive.factories.HiveCatalogFactoryOptions;
-import org.apache.flink.types.Row;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -289,154 +274,6 @@ public class FlinkHiveCatalogIT extends FlinkCommonIT {
     Assertions.assertFalse(metalake.catalogExists(catalogName));
     Assertions.assertEquals(
         numCatalogs, tableEnv.listCatalogs().length, "The created catalog should be dropped.");
-  }
-
-  @Test
-  public void testCreateNoPartitionTable() {
-    String databaseName = "test_create_no_partition_table_db";
-    String tableName = "test_create_no_partition_table";
-    String comment = "test comment";
-    String key = "test key";
-    String value = "test value";
-
-    // 1. The NOT NULL constraint for column is only supported since Hive 3.0,
-    // but the current Gravitino Hive catalog only supports Hive 2.x.
-    // 2. Hive doesn't support Time and Timestamp with timezone type.
-    // 3. Flink SQL only support to create Interval Month and Second(3).
-    doWithSchema(
-        hiveCatalog,
-        databaseName,
-        catalog -> {
-          TableResult result =
-              sql(
-                  "CREATE TABLE %s "
-                      + "(string_type STRING COMMENT 'string_type', "
-                      + " double_type DOUBLE COMMENT 'double_type')"
-                      + " COMMENT '%s' WITH ("
-                      + "'%s' = '%s')",
-                  tableName, comment, key, value);
-          TestUtils.assertTableResult(result, ResultKind.SUCCESS);
-
-          Table table =
-              catalog.asTableCatalog().loadTable(NameIdentifier.of(databaseName, tableName));
-          Assertions.assertNotNull(table);
-          Assertions.assertEquals(comment, table.comment());
-          Assertions.assertEquals(value, table.properties().get(key));
-          Column[] columns =
-              new Column[] {
-                Column.of("string_type", Types.StringType.get(), "string_type", true, false, null),
-                Column.of("double_type", Types.DoubleType.get(), "double_type")
-              };
-          assertColumns(columns, table.columns());
-          Assertions.assertArrayEquals(EMPTY_TRANSFORM, table.partitioning());
-        },
-        true);
-  }
-
-  @Test
-  public void testListTables() {
-    String newSchema = "test_list_table_catalog";
-    Column[] columns = new Column[] {Column.of("user_id", Types.IntegerType.get(), "USER_ID")};
-    doWithSchema(
-        hiveCatalog,
-        newSchema,
-        catalog -> {
-          catalog
-              .asTableCatalog()
-              .createTable(
-                  NameIdentifier.of(newSchema, "test_table1"),
-                  columns,
-                  "comment1",
-                  ImmutableMap.of());
-          catalog
-              .asTableCatalog()
-              .createTable(
-                  NameIdentifier.of(newSchema, "test_table2"),
-                  columns,
-                  "comment2",
-                  ImmutableMap.of());
-          TableResult result = sql("SHOW TABLES");
-          TestUtils.assertTableResult(
-              result,
-              ResultKind.SUCCESS_WITH_CONTENT,
-              Row.of("test_table1"),
-              Row.of("test_table2"));
-        },
-        true);
-  }
-
-  @Test
-  public void testDropTable() {
-    String databaseName = "test_drop_table_db";
-    doWithSchema(
-        hiveCatalog,
-        databaseName,
-        catalog -> {
-          String tableName = "test_drop_table";
-          Column[] columns =
-              new Column[] {Column.of("user_id", Types.IntegerType.get(), "USER_ID")};
-          NameIdentifier identifier = NameIdentifier.of(databaseName, tableName);
-          catalog.asTableCatalog().createTable(identifier, columns, "comment1", ImmutableMap.of());
-          Assertions.assertTrue(catalog.asTableCatalog().tableExists(identifier));
-
-          TableResult result = sql("DROP TABLE %s", tableName);
-          TestUtils.assertTableResult(result, ResultKind.SUCCESS);
-          Assertions.assertFalse(catalog.asTableCatalog().tableExists(identifier));
-        },
-        true);
-  }
-
-  @Test
-  public void testGetTable() {
-    String databaseName = "test_get_hive_table";
-    Column[] columns =
-        new Column[] {
-          Column.of("string_type", Types.StringType.get(), "string_type", true, false, null),
-          Column.of("double_type", Types.DoubleType.get(), "double_type")
-        };
-
-    doWithSchema(
-        hiveCatalog,
-        databaseName,
-        catalog -> {
-          String tableName = "test_desc_table";
-          String comment = "comment1";
-          catalog
-              .asTableCatalog()
-              .createTable(
-                  NameIdentifier.of(databaseName, "test_desc_table"),
-                  columns,
-                  comment,
-                  ImmutableMap.of("k1", "v1"));
-
-          Optional<org.apache.flink.table.catalog.Catalog> flinkCatalog =
-              tableEnv.getCatalog(DEFAULT_HIVE_CATALOG);
-          Assertions.assertTrue(flinkCatalog.isPresent());
-          try {
-            CatalogBaseTable table =
-                flinkCatalog.get().getTable(new ObjectPath(databaseName, tableName));
-            Assertions.assertNotNull(table);
-            Assertions.assertEquals(CatalogBaseTable.TableKind.TABLE, table.getTableKind());
-            Assertions.assertEquals(comment, table.getComment());
-
-            org.apache.flink.table.catalog.Column[] expected =
-                new org.apache.flink.table.catalog.Column[] {
-                  org.apache.flink.table.catalog.Column.physical("string_type", DataTypes.STRING())
-                      .withComment("string_type"),
-                  org.apache.flink.table.catalog.Column.physical("double_type", DataTypes.DOUBLE())
-                      .withComment("double_type")
-                };
-            org.apache.flink.table.catalog.Column[] actual =
-                toFlinkPhysicalColumn(table.getUnresolvedSchema().getColumns());
-            Assertions.assertArrayEquals(expected, actual);
-
-            CatalogTable catalogTable = (CatalogTable) table;
-            Assertions.assertFalse(catalogTable.isPartitioned());
-          } catch (TableNotExistException e) {
-            Assertions.fail(e);
-          }
-        },
-        true);
   }
 
   @Override
